Modules can now signal that they are system modules.
These modules receive every message at least once regardless
if they provide a matching command or not. These type
of module is needed since there are modules which are
quite and do not need to be called via a command (e.g.
website-title-grabbing).
The URL and ICD modules have been refactored to use be
a system module.

Signed-off-by: Kristoffer Schneider <kschneider@codingfalx.de>